requirements.txt
Django>=4.2,<5
djangorestframework
django-filter
djangorestframework-simplejwt
python-dotenv
django-cors-headers
drf-yasg
whitenoise>=6.6.0
psycopg[binary]
gunicorn
whitenoise
pytest
pytest-django
pytest-cov
factory_boy
Faker
